requirements.txt
boto3==1.42.89
botocore==1.42.89
cffi==2.0.0
cryptography==46.0.7
jmespath==1.1.0
numpy==2.2.6
pandas==2.3.3
py4j==0.10.9.9
pycparser==3.0
PyMySQL==1.1.2
pyspark==4.1.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
